% Auto-generated: do not edit by hand
\name{rDashWebsocket}

\alias{rDashWebsocket}

\title{DashWebsocket component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
rDashWebsocket(id=NULL, msg=NULL, url=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{msg}{Character. The websocket response message.}

\item{url}{Character. The url for websocket.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

